# Use JsValue for property access