--- a/Jint.Tests.Test262/GlobalTests.cs
+++ b/Jint.Tests.Test262/GlobalTests.cs
@@ -1,0 +1,15 @@
+﻿using Xunit;
+
+namespace Jint.Tests.Test262
+{
+    public class GlobalTests : Test262Test
+    {
+        [Theory(DisplayName = "built-ins\\global")]
+        [MemberData(nameof(SourceFiles), "built-ins\\global", false)]
+        [MemberData(nameof(SourceFiles), "built-ins\\global", true, Skip = "Skipped")]
+        protected void RunTest(SourceFile sourceFile)
+        {
+            RunTestInternal(sourceFile);
+        }
+    }
+}