--- a/Jint/Collections/PropertyDictionary.cs
+++ b/Jint/Collections/PropertyDictionary.cs
@@ -1,15 +1,14 @@
-using System.Collections.Generic;
 using Jint.Runtime.Descriptors;
 
 namespace Jint.Collections
 {
-    internal sealed class PropertyDictionary : HybridDictionary<string, PropertyDescriptor>
+    internal sealed class PropertyDictionary : HybridDictionary<PropertyDescriptor>
     {
         public PropertyDictionary()
         {
         }
 
-        public PropertyDictionary(int capacity) : base(capacity)
+        public PropertyDictionary(int capacity, bool checkExistingKeys) : base(capacity, checkExistingKeys)
         {
         }
     }