--- a/Jint/Collections/SymbolDictionary.cs
+++ b/Jint/Collections/SymbolDictionary.cs
@@ -1,0 +1,16 @@
+using Jint.Native;
+using Jint.Runtime.Descriptors;
+
+namespace Jint.Collections
+{
+    internal sealed class SymbolDictionary : DictionarySlim<JsSymbol, PropertyDescriptor>
+    {
+        public SymbolDictionary()
+        {
+        }
+
+        public SymbolDictionary(int capacity) : base(capacity)
+        {
+        }
+    }
+}