Using JsValue for property access aligns better with spec and later opens us for using them directly as keys in dictionary. This will reduce back and forth transformation of keys to string and JsValue. This should not affect normal usage as there are automatic conversions between common CLR types and JsValues.