@lahma 2020-02-02
Perf numbers after refactoring

## Esprima.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|3d-cube|470.2 ms|1.49 ms|12000.0000|1000.0000|0.0000|50.28 MB|
| **New** |	|	| **467.3 ms (-1%)** | **1.65 ms** | **12000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **50.94 MB (+1%)** |
| Old |Run|3d-morph|391.2 ms|2.39 ms|11000.0000|2000.0000|0.0000|68.5 MB|
| **New** |	|	| **395.3 ms (+1%)** | **2.88 ms** | **10000.0000 (-9%)** | **2000.0000 (0%)** | **0.0000** | **63.38 MB (-7%)** |
| Old |Run|3d-raytrace|338.3 ms|3.09 ms|21000.0000|3000.0000|0.0000|87.35 MB|
| **New** |	|	| **330.0 ms (-2%)** | **2.21 ms** | **20000.0000 (-5%)** | **3000.0000 (0%)** | **0.0000** | **84.95 MB (-3%)** |
| Old |Run|access-binary-trees|153.9 ms|0.50 ms|15750.0000|500.0000|0.0000|65.03 MB|
| **New** |	|	| **171.6 ms (+12%)** | **1.11 ms** | **14666.6667 (-7%)** | **333.3333 (-33%)** | **0.0000** | **60.86 MB (-6%)** |
| Old |Run|access-fannkuch|1,129.2 ms|15.34 ms|9000.0000|0.0000|0.0000|37.91 MB|
| **New** |	|	| **1,126.0 ms (0%)** | **10.05 ms** | **9000.0000 (0%)** | **0.0000** | **0.0000** | **37.92 MB (0%)** |
| Old |Run|access-nbody|378.5 ms|1.16 ms|15000.0000|0.0000|0.0000|62.28 MB|
| **New** |	|	| **426.3 ms (+13%)** | **1.43 ms** | **15000.0000 (0%)** | **0.0000** | **0.0000** | **61.97 MB (0%)** |
| Old |Run|access-nsieve|487.6 ms|5.37 ms|11000.0000|3000.0000|0.0000|65.98 MB|
| **New** |	|	| **424.3 ms (-13%)** | **1.56 ms** | **10000.0000 (-9%)** | **4000.0000 (+33%)** | **1000.0000** | **55.57 MB (-16%)** |
| Old |Run|bitop(...)-byte [24]|299.2 ms|2.02 ms|13000.0000|0.0000|0.0000|52.7 MB|
| **New** |	|	| **291.7 ms (-3%)** | **0.91 ms** | **13000.0000 (0%)** | **0.0000** | **0.0000** | **53.68 MB (+2%)** |
| Old |Run|bitops-bits-in-byte|505.2 ms|2.17 ms|8000.0000|0.0000|0.0000|34.14 MB|
| **New** |	|	| **489.0 ms (-3%)** | **2.15 ms** | **8000.0000 (0%)** | **0.0000** | **0.0000** | **34.82 MB (+2%)** |
| Old |Run|bitops-bitwise-and|321.8 ms|1.43 ms|13000.0000|0.0000|0.0000|54.63 MB|
| **New** |	|	| **428.5 ms (+33%)** | **1.15 ms** | **13000.0000 (0%)** | **0.0000** | **0.0000** | **54.64 MB (0%)** |
| Old |Run|bitops-nsieve-bits|425.8 ms|3.09 ms|19000.0000|1000.0000|0.0000|80.65 MB|
| **New** |	|	| **417.4 ms (-2%)** | **2.50 ms** | **16000.0000 (-16%)** | **1000.0000 (0%)** | **0.0000** | **66.48 MB (-18%)** |
| Old |Run|contr(...)rsive [21]|249.3 ms|4.19 ms|19000.0000|0.0000|0.0000|78.16 MB|
| **New** |	|	| **232.3 ms (-7%)** | **1.50 ms** | **20000.0000 (+5%)** | **2000.0000** | **0.0000** | **79.78 MB (+2%)** |
| Old |Run|crypto-aes|337.7 ms|1.04 ms|5000.0000|1000.0000|0.0000|23.5 MB|
| **New** |	|	| **317.5 ms (-6%)** | **1.60 ms** | **6000.0000 (+20%)** | **1000.0000 (0%)** | **0.0000** | **27.04 MB (+15%)** |
| Old |Run|crypto-md5|229.3 ms|1.26 ms|19000.0000|1000.0000|0.0000|79.53 MB|
| **New** |	|	| **213.7 ms (-7%)** | **2.62 ms** | **18000.0000 (-5%)** | **1000.0000 (0%)** | **0.0000** | **75.51 MB (-5%)** |
| Old |Run|crypto-sha1|210.0 ms|1.52 ms|15000.0000|2000.0000|0.0000|64.14 MB|
| **New** |	|	| **211.7 ms (+1%)** | **1.20 ms** | **15000.0000 (0%)** | **1000.0000 (-50%)** | **0.0000** | **62.4 MB (-3%)** |
| Old |Run|date-format-tofte|232.3 ms|1.10 ms|15000.0000|1000.0000|0.0000|60.99 MB|
| **New** |	|	| **223.5 ms (-4%)** | **0.92 ms** | **15000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **60.76 MB (0%)** |
| Old |Run|date-format-xparb|112.5 ms|0.63 ms|5600.0000|200.0000|0.0000|23.82 MB|
| **New** |	|	| **121.7 ms (+8%)** | **0.27 ms** | **5600.0000 (0%)** | **200.0000 (0%)** | **0.0000** | **23.58 MB (-1%)** |
| Old |Run|math-cordic|686.3 ms|1.48 ms|24000.0000|0.0000|0.0000|98.78 MB|
| **New** |	|	| **715.1 ms (+4%)** | **0.80 ms** | **24000.0000 (0%)** | **0.0000** | **0.0000** | **97.83 MB (-1%)** |
| Old |Run|math-partial-sums|308.0 ms|1.02 ms|9000.0000|0.0000|0.0000|38.29 MB|
| **New** |	|	| **311.7 ms (+1%)** | **1.42 ms** | **9000.0000 (0%)** | **0.0000** | **0.0000** | **38.29 MB (0%)** |
| Old |Run|math-spectral-norm|270.4 ms|2.53 ms|12000.0000|0.0000|0.0000|47.98 MB|
| **New** |	|	| **228.7 ms (-15%)** | **1.30 ms** | **12000.0000 (0%)** | **0.0000** | **0.0000** | **48.93 MB (+2%)** |
| Old |Run|regexp-dna|333.2 ms|1.69 ms|0.0000|0.0000|0.0000|18.25 MB|
| **New** |	|	| **335.4 ms (+1%)** | **1.10 ms** | **0.0000** | **0.0000** | **0.0000** | **18.25 MB (0%)** |
| Old |Run|string-base64|183.2 ms|0.61 ms|8000.0000|1000.0000|0.0000|35.07 MB|
| **New** |	|	| **171.1 ms (-7%)** | **1.23 ms** | **1000.0000 (-88%)** | **0.0000 (-100%)** | **0.0000** | **5.44 MB (-84%)** |
| Old |Run|string-fasta|413.4 ms|3.87 ms|25000.0000|0.0000|0.0000|102.85 MB|
| **New** |	|	| **392.7 ms (-5%)** | **1.98 ms** | **27000.0000 (+8%)** | **0.0000** | **0.0000** | **109.2 MB (+6%)** |
| Old |Run|string-tagcloud|156.3 ms|0.66 ms|10000.0000|2000.0000|0.0000|52.01 MB|
| **New** |	|	| **158.6 ms (+1%)** | **2.03 ms** | **8000.0000 (-20%)** | **1000.0000 (-50%)** | **0.0000** | **44.89 MB (-14%)** |
| Old |Run|string-unpack-code|174.5 ms|1.36 ms|15000.0000|4000.0000|1000.0000|70.79 MB|
| **New** |	|	| **167.9 ms (-4%)** | **2.00 ms** | **15000.0000 (0%)** | **5000.0000 (+25%)** | **1000.0000 (0%)** | **71.59 MB (+1%)** |
| Old |Run|strin(...)input [21]|1,085.4 ms|13.47 ms|1152000.0000|1137000.0000|1136000.0000|6190.99 MB|
| **New** |	|	| **1,092.1 ms (+1%)** | **11.19 ms** | **1063000.0000 (-8%)** | **1049000.0000 (-8%)** | **1049000.0000 (-8%)** | **6182.43 MB (0%)** |


## Jint.Benchmark.ArrayBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Slice|100|372.4 us|1.18 us|172.8516|0.0000|0.0000|707.03 KB|
| **New** |	|	| **396.9 us (+7%)** | **2.61 us** | **173.8281 (+1%)** | **0.0000** | **0.0000** | **710.16 KB (0%)** |
| Old |Concat|100|511.0 us|1.40 us|188.4766|0.0000|0.0000|771.88 KB|
| **New** |	|	| **585.0 us (+14%)** | **1.59 us** | **207.0313 (+10%)** | **0.0000** | **0.0000** | **847.66 KB (+10%)** |
| Old |Unshift|100|16,159.0 us|35.05 us|3687.5000|0.0000|0.0000|15078.13 KB|
| **New** |	|	| **17,642.5 us (+9%)** | **81.87 us** | **3687.5000 (0%)** | **0.0000** | **0.0000** | **15092.19 KB (0%)** |
| Old |Push|100|8,260.2 us|16.65 us|453.1250|0.0000|0.0000|1867.97 KB|
| **New** |	|	| **9,597.4 us (+16%)** | **21.83 us** | **453.1250 (0%)** | **0.0000** | **0.0000** | **1888.28 KB (+1%)** |
| Old |Index|100|9,393.8 us|46.68 us|562.5000|0.0000|0.0000|2345.31 KB|
| **New** |	|	| **10,550.1 us (+12%)** | **47.96 us** | **640.6250 (+14%)** | **0.0000** | **0.0000** | **2624.22 KB (+12%)** |
| Old |Map|100|3,607.1 us|7.17 us|753.9063|0.0000|0.0000|3093.79 KB|
| **New** |	|	| **3,645.7 us (+1%)** | **17.56 us** | **773.4375 (+3%)** | **0.0000** | **0.0000** | **3171.88 KB (+3%)** |
| Old |Apply|100|576.9 us|2.50 us|202.1484|0.0000|0.0000|828.13 KB|
| **New** |	|	| **603.1 us (+5%)** | **4.35 us** | **204.1016 (+1%)** | **0.0000** | **0.0000** | **834.38 KB (+1%)** |
| Old |JsonStringifyParse|100|4,510.0 us|18.97 us|1226.5625|0.0000|0.0000|5038.28 KB|
| **New** |	|	| **4,217.6 us (-6%)** | **23.54 us** | **1289.0625 (+5%)** | **0.0000** | **0.0000** | **5268.76 KB (+5%)** |
| Old |FilterWithString|100|39,463.7 us|250.81 us|6307.6923|0.0000|0.0000|25767.19 KB|
| **New** |	|	| **40,167.4 us (+2%)** | **374.63 us** | **6461.5385 (+2%)** | **0.0000** | **0.0000** | **26557.03 KB (+3%)** |


## Jint.Benchmark.ArrayStressBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Jint|20|526.9 ms|1.47 ms|60000.0000|13000.0000|0.0000|265.26 MB|
| **New** |	|	| **577.6 ms (+10%)** | **1.01 ms** | **60000.0000 (0%)** | **10000.0000 (-23%)** | **0.0000** | **259.8 MB (-2%)** |

## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|59.52 ms|0.213 ms|1444.4444|444.4444|0.0000|8094.98 KB|
| **New** |	|	| **55.51 ms (-7%)** | **0.115 ms** | **1500.0000 (+4%)** | **500.0000 (+13%)** | **0.0000** | **8242.87 KB (+2%)** |
| Old |Run|dromaeo-core-eval|14.38 ms|0.090 ms|93.7500|15.6250|0.0000|443.15 KB|
| **New** |	|	| **15.61 ms (+9%)** | **0.057 ms** | **93.7500 (0%)** | **0.0000 (-100%)** | **0.0000** | **450.94 KB (+2%)** |
| Old |Run|dromaeo-object-array|166.79 ms|1.759 ms|41000.0000|1000.0000|0.0000|176108.92 KB|
| **New** |	|	| **166.88 ms (0%)** | **1.087 ms** | **41000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **176071.13 KB (0%)** |
| Old |Run|droma(...)tring [21]|582.99 ms|6.598 ms|112000.0000|70000.0000|68000.0000|1378875.84 KB|
| **New** |	|	| **551.34 ms (-5%)** | **5.930 ms** | **105000.0000 (-6%)** | **71000.0000 (+1%)** | **70000.0000 (+3%)** | **1342669.25 KB (-3%)** |
| Old |Run|droma(...)ase64 [21]|122.70 ms|0.804 ms|5800.0000|200.0000|0.0000|24428.98 KB|
| **New** |	|	| **110.83 ms (-10%)** | **0.280 ms** | **800.0000 (-86%)** | **200.0000 (0%)** | **0.0000** | **3864.88 KB (-84%)** |


## Jint.Benchmark.EngineConstructionBenchmark

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |BuildEngine|2.594 us|0.0156 us|2.3727|0.0000|0.0000|9.7 KB|
| **New** |	| **2.549 us (-2%)** | **0.0098 us** | **2.4414 (+3%)** | **0.0000** | **0.0000** | **9.98 KB (+3%)** |


## Jint.Benchmark.EvaluationBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Jint|20|905.7 us|2.86 us|141.6016|0.0000|0.0000|581.11 KB|
| **New** |	|	| **885.1 us (-2%)** | **1.97 us** | **143.5547 (+1%)** | **0.0000** | **0.0000** | **588.91 KB (+1%)** |


## Jint.Benchmark.InteropBenchmark

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |DelegatesCanBeSet|7.931 us|0.0428 us|1.4219|0.0000|0.0000|6033 B|
| **New** |	| **7.667 us (-3%)** | **0.0406 us** | **1.4609 (+3%)** | **0.0000** | **0.0000** | **6136 B (+2%)** |
| Old |ExtraParametersAreIgnored|12.152 us|0.0360 us|1.9688|0.0000|0.0000|8281 B|
| **New** |	| **11.810 us (-3%)** | **0.0531 us** | **2.0000 (+2%)** | **0.0000** | **0.0000** | **8401 B (+1%)** |
| Old |GetObjectProperties|6.236 us|0.0233 us|1.3125|0.0000|0.0000|5504 B|
| **New** |	| **6.013 us (-4%)** | **0.0273 us** | **1.3203 (+1%)** | **0.0000** | **0.0000** | **5528 B (0%)** |
| Old |InvokeObjectMethods|6.960 us|0.0132 us|1.5469|0.0000|0.0000|6480 B|
| **New** |	| **7.146 us (+3%)** | **0.0311 us** | **1.5547 (+1%)** | **0.0000** | **0.0000** | **6512 B (0%)** |
| Old |SetObjectProperties|8.899 us|0.0480 us|1.7344|0.0000|0.0000|7264 B|
| **New** |	| **9.308 us (+5%)** | **0.0192 us** | **1.7344 (0%)** | **0.0000** | **0.0000** | **7313 B (+1%)** |
| Old |GetIndexUsingStringKey|18.769 us|0.3147 us|2.9375|0.0000|0.0000|12385 B|
| **New** |	| **18.824 us (0%)** | **0.0867 us** | **2.9063 (-1%)** | **0.0000** | **0.0000** | **12232 B (-1%)** |
| Old |GenericMethods|20.953 us|0.0724 us|2.9688|0.0000|0.0000|12505 B|
| **New** |	| **20.052 us (-4%)** | **0.0616 us** | **3.0000 (+1%)** | **0.0000** | **0.0000** | **12569 B (+1%)** |
| Old |MultiGenericTypes|44.379 us|0.2392 us|5.6667|0.0000|0.0000|23706 B|
| **New** |	| **44.415 us (0%)** | **0.1039 us** | **5.6667 (0%)** | **0.0000** | **0.0000** | **24018 B (+1%)** |
| Old |IndexOnList|12.928 us|0.0711 us|2.0313|0.0000|0.0000|8497 B|
| **New** |	| **13.553 us (+5%)** | **0.0761 us** | **2.0469 (+1%)** | **0.0000** | **0.0000** | **8600 B (+1%)** |
| Old |EcmaValuesAreAutomaticallyConvertedWhenSetInPoco|NA|NA|-|-|-|-|
| **New** |	| **NA** | **NA** | **-** | **-** | **-** | **-** |
| Old |Trim|15.970 us|0.1121 us|2.7500|0.0000|0.0000|11616 B|
| **New** |	| **16.061 us (+1%)** | **0.0381 us** | **2.7813 (+1%)** | **0.0000** | **0.0000** | **11633 B (0%)** |
| Old |MathFloor|NA|NA|-|-|-|-|
| **New** |	| **NA** | **NA** | **-** | **-** | **-** | **-** |
| Old |DelegateAsFunction|8.101 us|0.0475 us|1.4063|0.0000|0.0000|5920 B|
| **New** |	| **7.649 us (-6%)** | **0.0315 us** | **1.4375 (+2%)** | **0.0000** | **0.0000** | **6024 B (+2%)** |
| Old |ConvertArrayToArrayInstance|13.663 us|0.0680 us|2.4844|0.0000|0.0000|10448 B|
| **New** |	| **13.850 us (+1%)** | **0.0745 us** | **2.5313 (+2%)** | **0.0000** | **0.0000** | **10624 B (+2%)** |
| Old |ConvertListsToArrayInstance|16.369 us|0.1173 us|2.8438|0.0000|0.0000|11912 B|
| **New** |	| **16.621 us (+2%)** | **0.1390 us** | **2.8750 (+1%)** | **0.0000** | **0.0000** | **12136 B (+2%)** |
| Old |ConvertArrayInstanceToArray|3.699 us|0.0329 us|1.1563|0.0000|0.0000|4856 B|
| **New** |	| **3.483 us (-6%)** | **0.0081 us** | **1.1172 (-3%)** | **0.0000** | **0.0000** | **4680 B (-4%)** |
| Old |LoopWithNativeEnumerator|7.803 us|0.0514 us|1.4375|0.0000|0.0000|6048 B|
| **New** |	| **7.475 us (-4%)** | **0.0138 us** | **1.4609 (+2%)** | **0.0000** | **0.0000** | **6128 B (+1%)** |


## Jint.Benchmark.LinqJsBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Jint|10|62.64 ms|0.184 ms|5000.0000|2375.0000|0.0000|29.87 MB|
| **New** |	|	| **61.92 ms (-1%)** | **0.326 ms** | **5000.0000 (0%)** | **2333.3333 (-2%)** | **0.0000** | **29.92 MB (0%)** |


## Jint.Benchmark.MinimalScriptBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Jint|10|25.87 us|0.178 us|8.8501|0.0000|0.0000|36.25 KB|
| **New** |	|	| **26.24 us (+1%)** | **0.034 us** | **9.0942 (+3%)** | **0.0000** | **0.0000** | **37.19 KB (+3%)** |


## Jint.Benchmark.StopwatchBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Jint|1|975.7 ms|5.31 ms|29000.0000|0.0000|0.0000|118.74 MB|
| **New** |	|	| **1.146 s (+20%)** | **0.0026 s** | **30000.0000 (+3%)** | **0.0000** | **0.0000** | **122.46 MB (+3%)** |


## Jint.Benchmark.StringBuilderBenchmark

| **Diff**|Method|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |One|2.203 us|0.0136 us|0.8583|0.0000|0.0000|3.52 KB|
| **New** |	| **2.202 us (0%)** | **0.0080 us** | **0.8698 (+1%)** | **0.0000** | **0.0000** | **3.55 KB (+1%)** |
| Old |Two|2.930 us|0.0198 us|0.9995|0.0000|0.0000|4.09 KB|
| **New** |	| **2.897 us (-1%)** | **0.0119 us** | **1.0147 (+2%)** | **0.0000** | **0.0000** | **4.16 KB (+2%)** |
| Old |Three|3.618 us|0.0161 us|1.1406|0.0000|0.0000|4.66 KB|
| **New** |	| **3.587 us (-1%)** | **0.0071 us** | **1.1597 (+2%)** | **0.0000** | **0.0000** | **4.76 KB (+2%)** |


## Jint.Benchmark.UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|248.1 ms|1.82 ms|43000.0000|11000.0000|0.0000|171.78 MB|
| **New** |	|	| **271.2 ms (+9%)** | **0.95 ms** | **44000.0000 (+2%)** | **11000.0000 (0%)** | **0.0000** | **178 MB (+4%)** |
| Old |Benchmark|500|194.8 ms|1.15 ms|37333.3333|2000.0000|0.0000|149.96 MB|
| **New** |	|	| **205.5 ms (+5%)** | **0.55 ms** | **38666.6667 (+4%)** | **2333.3333 (+17%)** | **0.0000** | **156.16 MB (+4%)** |